pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dco"
version = "0.0.1"
description = "Dynamic Config Orchestrator - layered typed configuration loader"
authors = [{name = "Safwan Bakkar", email = "safwanbakkar.dev@hotmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "pydantic>=1.10",
    "python-dotenv>=0.21",
    "PyYAML>=6.0",
    "watchdog>=3.0",
    "boto3>=1.26",
]

[project.scripts]
dco = "dco.cli:main"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
extend-ignore = ["E203"]  # if you want to suppress specific rules
